Fix build-*llvm.sh to build either mclinker 2.4.0 or the coming 2.5.0
Differences between 2.4.0 and 2.5.0:
1. binary location is different
2. Use ld.lite in 2.5.0 (1/10 size of 2.4.0) and rename it to ld.mcld

Also misc fixes:
1. Build llvm-as as well
2. One typo about -z

+++ build/tools/build-device-llvm.sh
@@ -238,7 +238,12 @@ for abi in $ABIS; do
   fail_panic "Couldn't compile mclinker"
 
   run mkdir -p $TOOLCHAIN_BUILD_PREFIX/$abi
-  run cp -f $MCLINKER_BUILD_OUT/optimized/ld.mcld $TOOLCHAIN_BUILD_PREFIX/$abi
+  if [ -f $MCLINKER_BUILD_OUT/tools/ld.lite/ld.lite ]; then
+    run cp -f $MCLINKER_BUILD_OUT/tools/ld.lite/ld.lite $TOOLCHAIN_BUILD_PREFIX/$abi/ld.mcld
+  else
+    run cp -f $MCLINKER_BUILD_OUT/optimized/ld.mcld $TOOLCHAIN_BUILD_PREFIX/$abi
+  fi
+  fail_panic "Couldn't copy mclinker"
 
   # Strip
   STRIP=$BUILD_OUT/ndk-standalone-$arch/bin/$toolchain_prefix-strip
